backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==2.2.2
opencv-python-headless==4.11.0.86
ultralytics>=8.3.100
slowapi>=0.1.9

httpx==0.27.2
python-dotenv
